extracted AuthorizedStorageAccounts non storage-specific logic to AuthorizedAccount
