+++ addon_service/addon_operation_invocation/models.py
@@ -87,7 +87,7 @@ class AddonOperationInvocation(AddonsServiceBaseModel):
     def storage_imp_config(self) -> StorageConfig:
         if self.thru_addon:
             return self.thru_addon.storage_imp_config()
-        return self.thru_account.storage_imp_config()
+        return self.thru_account.storage_imp_config
 
     def set_exception(self, exception: BaseException) -> None:
         self.invocation_status = InvocationStatus.ERROR
