+++ addon_service/authorized_storage_account/views.py
@@ -1,32 +1,8 @@
-from addon_service.common.permissions import (
-    IsAuthenticated,
-    SessionUserIsOwner,
-)
-from addon_service.common.viewsets import RetrieveWriteDeleteViewSet
-
+from ..authorized_account.views import AuthorizedAccountViewSet
 from .models import AuthorizedStorageAccount
 from .serializers import AuthorizedStorageAccountSerializer
 
 
-class AuthorizedStorageAccountViewSet(RetrieveWriteDeleteViewSet):
+class AuthorizedStorageAccountViewSet(AuthorizedAccountViewSet):
     queryset = AuthorizedStorageAccount.objects.all()
     serializer_class = AuthorizedStorageAccountSerializer
-
-    def get_permissions(self):
-        match self.action:
-            case (
-                "retrieve"
-                | "retrieve_related"
-                | "partial_update"
-                | "update"
-                | "destroy"
-            ):
-                return [IsAuthenticated(), SessionUserIsOwner()]
-            case "create":
-                return [IsAuthenticated()]
-            case None:
-                return super().get_permissions()
-            case _:
-                raise NotImplementedError(
-                    f"no permission implemented for action '{self.action}'"
-                )
